pyproject.toml
[tool.poetry]
name = "repl_python3_BitcoinBot"
version = "0.1.0"
description = ""
authors = ["AntoineButeau <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
